Refactor to create new `RemoteCommands` class
