+++ test/data/test-command-file-1.txt
@@ -1,11 +1,15 @@
 # should be split into 2 commands
 LOAD file1.mp3 file2.mp3
+
 # should be added as one
 LOAD file3.mp3
-FROM_FILE command_file.txt
-CLUSTER unclustered
-  FINGERPRINT  
+
+# should be ignored because circular reference
+FROM_FILE test/data/test-command-file-1.txt
+
 # should be ignored
  
 
 #commented command
+
+FROM_FILE test/data/test-command-file-2.txt
